machine/atmega328pb: refactor to enable extra uart
