+++ src/machine/board_arduino.go
@@ -2,11 +2,6 @@
 
 package machine
 
-// Return the current CPU frequency in hertz.
-func CPUFrequency() uint32 {
-	return 16000000
-}
-
 // Digital pins, marked as plain numbers on the board.
 const (
 	D0  = PD0 // RX
