+++ targets/atmega328pb.json
@@ -1,0 +1,15 @@
+{
+	"inherits": ["avr"],
+	"cpu": "atmega328pb",
+	"build-tags": ["atmega328pb", "atmega", "avr5"],
+	"ldflags": [
+		"--defsym=_bootloader_size=512",
+		"--defsym=_stack_size=512"
+	],
+	"serial": "uart",
+	"linkerscript": "src/device/avr/atmega328pb.ld",
+	"extra-files": [
+		"targets/avr.S",
+		"src/device/avr/atmega328pb.s"
+	]
+}
